+++ stac_fastapi/core/setup.py
@@ -1,4 +1,4 @@
-"""stac_fastapi: elasticsearch module."""
+"""stac_fastapi: core elasticsearch/ opensearch module."""
 
 from setuptools import find_namespace_packages, setup
 
@@ -13,30 +13,12 @@ install_requires = [
     "stac-fastapi.types==2.4.9",
     "stac-fastapi.api==2.4.9",
     "stac-fastapi.extensions==2.4.9",
-    # "elasticsearch[async]==8.11.0",
-    # "elasticsearch-dsl==8.11.0",
     "pystac[validation]",
-    # "uvicorn",
     "orjson",
     "overrides",
-    # "starlette",
     "geojson-pydantic",
     "pygeofilter==0.2.1",
 ]
-
-# extra_reqs = {
-#     "dev": [
-#         "pytest",
-#         "pytest-cov",
-#         "pytest-asyncio",
-#         "pre-commit",
-#         "requests",
-#         "ciso8601",
-#         "httpx",
-#     ],
-#     "docs": ["mkdocs", "mkdocs-material", "pdocs"],
-#     "server": ["uvicorn[standard]==0.19.0"],
-# }
 
 setup(
     name="stac-fastapi.core",
